+++ Runtime/Libraries/NUtils.cs
@@ -20,5 +20,14 @@ namespace Nazio_LT.Tools.Core
 
         /// <summary> Return the last element of a List.</summary>
         public static T Last<T>(this List<T> _list) => _list[_list.Count - 1];
+
+        public static T[] Merge<T>(T[] _array1, T[] _array2)
+        {
+            T[] _result = new T[_array1.Length + _array2.Length];
+            _array1.CopyTo(_result, 0);
+            _array2.CopyTo(_result, _array1.Length);
+
+            return _result;
+        }
     }
 }