+++ Runtime/Models/InputBuffer.cs
@@ -3,8 +3,35 @@ using System;
 
 namespace Nazio_LT.Tools.Core
 {
+    public abstract class InputBuffer
+    {
+        protected float lastPressedTime = 999f;
+        protected bool canConsumme = false, consummed = true;
+
+        private const float TIME_TO_SAVE = 0.2f;
+
+        public abstract void TryExecute();
+
+        /// <summary>Reset the object.</summary>
+        public void Reset()
+        {
+            canConsumme = false;
+            consummed = true;
+        }
+
+        /// <summary>Consume input if it's available and if the callback return true.</summary>
+        protected void ExecuteIfInputIsAvailable<T>(Func<T, bool> _callback, T _value)
+        {
+            if (!available) return;
+
+            if (_callback(_value)) consummed = true;
+        }
+
+        private bool available => !consummed && (canConsumme || (Time.time - lastPressedTime) < TIME_TO_SAVE);
+    }
+
     /// <summary>Save inputs.</summary>
-    public class InputBuffer<T>
+    public class InputBuffer<T> : InputBuffer
     {
         public InputBuffer(Func<T, bool> _linkedAction)
         {
@@ -12,18 +39,7 @@ namespace Nazio_LT.Tools.Core
         }
 
         private readonly Func<T, bool> linkedAction;
-
         private T value = default;
-        private float lastPressedTime = 999f;
-        private bool canConsumme = false, consummed = true;
-
-        private const float TIME_TO_SAVE = 0.2f;
-
-        public void Reset()
-        {
-            canConsumme = false;
-            consummed = true;
-        }
 
         public void Input(T _value, bool _performed)
         {
@@ -40,16 +56,10 @@ namespace Nazio_LT.Tools.Core
             consummed = false;
         }
 
-        /// <summary>Consume input if it's available and if the callback return true.</summary>
-        public void ExecuteIfInputIsAvailable(Func<T, bool> _callback)
-        {
-            if (!available) return;
+        #region Overrided Methods
 
-            if (_callback(value)) consummed = true;
-        }
+        public override void TryExecute() => ExecuteIfInputIsAvailable<T>(linkedAction, value);
 
-        public void TryExecute() => ExecuteIfInputIsAvailable(linkedAction);
-
-        private bool available => !consummed && (canConsumme || (Time.time - lastPressedTime) < TIME_TO_SAVE);
+        #endregion
     }
 }