+++ Runtime/Libraries/NUtils.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System;
 using UnityEngine;
 
 namespace Nazio_LT.Tools.Core
@@ -21,6 +22,7 @@ namespace Nazio_LT.Tools.Core
         /// <summary> Return the last element of a List.</summary>
         public static T Last<T>(this List<T> _list) => _list[_list.Count - 1];
 
+        /// <summary> Merge array 1 and 2.</summary>
         public static T[] Merge<T>(T[] _array1, T[] _array2)
         {
             T[] _result = new T[_array1.Length + _array2.Length];
@@ -28,6 +30,18 @@ namespace Nazio_LT.Tools.Core
             _array2.CopyTo(_result, _array1.Length);
 
             return _result;
+        }
+
+        /// <summary>Double for : for i { for j }</summary>
+        public static void DoubleFor(int _iLenght, int _jLeght, Action<int, int> _callbackAction)
+        {
+            for (int i = 0; i < _iLenght; i++)
+            {
+                for (int j = 0; j < _jLeght; j++)
+                {
+                    _callbackAction(i, j);
+                }
+            }
         }
     }
 }