+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Display If. ([]())
 * Adapt GUI. ([]())
 * Draw Header. ([]())
+* FoldableOutScriptableObject ([]())
 
 ## [1.1.0](https://github.com/NazioLT/NZCore/releases/tag/1.1.0) (2022-12-8)